Fix invisible positions after login (#33)
* Set minimum size of PortfolioPanel after adding positions.

This fixes #32 but more refactoring is to come.

* Set minimal size of panels.

It's not clear if this is actually doing anything. It's possible that our use of wxScrolledWindow to show positions allows all child components to expand to their minimal size naturally. However, most example code showing sizers uses SetSizerAndFit() so parents extend to contain their children.