Use oatpp code generation for API login/logout actions.
This is a fairly polished proof of concept but does require some cleanup that can only be done after the entire API surface is converted to oatpp. For now, we create a `_Api` class that is wrapped by the original `Api` class since we don't have a lot of insight into what code oatpp is generating.